Turn on 'treat warnings as errors' in sipXproxy so that printf() problems will be detected at compile time. Correct and/or improve things that are now flagged by the compiler.
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@18288 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXproxy/lib/authplugins/NatTraversalAgent/NatTraversalRules.cpp
@@ -248,7 +248,9 @@ void NatTraversalRules::initializeNatTraversalInfo( void )
             if( sscanf( tempString.data(), "%d:%d", &startPort, &endPort ) == 2 &&
                 startPort < endPort )
             {
-               mMaxMediaRelaySessions = ( endPort - startPort ) * NAT_TRAVERSAL_SHARE / PORTS_PER_MEDIA_RELAY_SESSIONS;
+               mMaxMediaRelaySessions =
+                  (int) (( endPort - startPort ) * NAT_TRAVERSAL_SHARE /
+                         PORTS_PER_MEDIA_RELAY_SESSIONS);
                if( mMaxMediaRelaySessions > 0 )
                {
                   isValid = true;
